Separate test suites in desktop/mobile folders and define device with the command line
