+++ package.json
@@ -47,7 +47,9 @@
     "test": "run-s test:**",
     "test:snapshots": "jest --testPathPattern=\"storyshots.test.js\" --silent",
     "test:snapshots:update": "jest --testPathPattern=\"storyshots.test.js\" -u --silent",
-    "test:e2e": "npm run build:app && testcafe chrome ./tests/e2e",
+    "test:e2e": "npm run build:app && npm run test:testcafe-desktop && npm run test:testcafe-mobile",
+    "test:testcafe-desktop": "testcafe chrome ./tests/e2e/desktop",
+    "test:testcafe-mobile": "testcafe \"chrome:emulation:mobile=true;device=iphone 6\" ./tests/e2e/mobile",
     "test:lighthouse": "npm run build:app && jest --testPathPattern=\"lighthouse.test.js\"",
     "develop": "npm start",
     "start": "gatsby develop",
