+++ tests/e2e/desktop/topNav-dropdown-test.js
@@ -1,4 +1,4 @@
-import createServer from '../createServer'
+import createServer from '../../createServer'
 import { Selector, ClientFunction } from 'testcafe'
 
 require('dotenv').config()
@@ -20,7 +20,7 @@ fixture`Top Nav Menu`.page`${baseUrl}`
 test('we are on the homepage', async t => {
   const location = await getWindowLocation()
   await t
-    .resizeWindow(1280, 720)
+    // .resizeWindow(1280, 720)
     .expect(location.href)
     .contains(baseUrl)
 })
